Re-organized some code and added comments
